# swaggerSpecification.yml
openapi: 3.0.0
servers:
  - url: localhost:8080
info:
  description: This is the API for the 2.4 Code Generation Bank application
  version: "2.0.1"
  title: Bank API
paths:
...
  /bankaccounts:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Accounts
      summary: Get all Bank Accounts
      parameters:
        - in: query
          name: userId
          description: Get all the accounts for a specific user
          schema:
            type: integer
      operationId: getAllAccounts
      description: Calling this allows you to fetch all Bank Accounts
      responses:
        "200":
          description: The Bank Accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArrayOfBankAccounts"
        "400":
          description: bad input parameter
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Accounts
      summary: Create a new BankAccount
      requestBody:
        description: Transaction object
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccount"
      operationId: createAccount
      description: Calling this allows you to create a sBank Account
      responses:
        "200":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
        "400":
          description: bad input parameter
  /bankaccount/{id}:
    get:
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: IBAN of the Bank Account to get
      tags:
        - Accounts
      summary: Get a Bank Account by IBAN
      operationId: getAccountById
      description: Calling this allows you to fetch a specific Bank Account by id
      responses:
        "200":
          description: the Bank Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
        "400":
          description: bad input parameter
    put:
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: IBAN of the Bank Account to update
      tags:
        - Accounts
      summary: Update a Bank Account by IBAN
      requestBody:
        description: BankAccount object
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccount"
      operationId: updateAccountById
      description: Calling this allows you to update a specific Bank Account by id
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
        "400":
          description: bad input parameter
    delete:
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: IBAN of the Bank Account to close
      tags:
        - Accounts
      summary: Closes a Bank Account by IBAN
      operationId: closeAccountById
      description: Calling this allows you to close a specific Bank Account by id
      responses:
        "200":
          description: The Bank account will be closed
        "400":
          description: bad input parameter
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BankAccount:
      type: object
      required:
        - userId
        - type
        - status
        - absoluteLimit
      properties:
        iban:
          description: The unique identifier for the account
          type: string
          pattern: ^[a-z]{2}[0-9]{2}[a-z0-9]{4}[0-9]{7}([a-z0-9]?){0,16}$
          example: NL01INHO0000000001
        userId:
          description: Owner of the bank account
          type: integer
          example: 1
        Status:
          description: Account status
          type: string
          enum: [ active, inactive ]
          example: active
        type:
          description: Account type
          type: string
          enum: [ regular, savings ]
          example: regular
        balance:
          description: The ammount of money on the account
          type: number
          format: double
          example: 500.50
        absoluteLimit:
          description: The absolute minimum of the account
          type: number
          format: double
          example: -1000
...
    ArrayOfBankAccounts:
      type: array
      items:
        $ref: "#/components/schemas/BankAccount"
...